backend/telegram-password-reset/requirements.txt
requests
bcrypt
psycopg2-binary